Level 2 initial dungeon room, parkour section added
This commit includes the first part of level 2. The first room is done, along with a short hall. The foundations of the first short parkour section is also complete. Lighting + fog effects are working as intended and out-of-bounds detection works. This level will be an introduction to the movement + combat mechanics I wrote last week (maybe more of a focus on movement... I find it more fun to mess with).